+++ star/app/_layout.tsx
@@ -16,14 +16,15 @@ export default function Layout() {
       }}
     >
       <Tab.Screen
+              name="Home"
+              component={HomeScreen}
+              options={{
+                tabBarIcon: ({ color, size }) => (
+                  <Icon name="home-outline" color={color} size={size} />
+                ), 
+              }}
+        
 
-        name="settings"
-          component={SettingsScreen}
-          options={{
-            tabBarIcon: ({ color, size }) => (
-              <Icon name="settings-outline" color={color} size={size} />
-            ),
-          }}
 
       />
 
@@ -38,14 +39,13 @@ export default function Layout() {
       />
 
       <Tab.Screen
-              name="Home"
-              component={HomeScreen}
-              options={{
-                tabBarIcon: ({ color, size }) => (
-                  <Icon name="home-outline" color={color} size={size} />
-                ), 
-              }}
-        
+        name="settings"
+        component={SettingsScreen}
+        options={{
+          tabBarIcon: ({ color, size }) => (
+            <Icon name="settings-outline" color={color} size={size} />
+          ),
+        }}
       />
     </Tab.Navigator>
   );
